jit: the runtime entry points JIT code calls move to their own file
src/builtin/jit_runtime.cpp is not a module - it is the set of extern "C"
symbols jitted and LLVM-AOT-emitted code links against, plus the address
takers the jit module binds. module_jit.cpp keeps the das module: bindings,
linker invocation and emit orchestration.
