+++ include/daScript/simulate/aot_builtin_jit.h
@@ -23,41 +23,5 @@ namespace das {
     bool das_remove_jit ( const Func func );
     bool das_instrument_jit ( void * pfun, const Func func, const LineInfo & info, Context & context );
     void * das_instrument_line_info ( const LineInfo & info, Context * context, LineInfoArg * at );
-    void * das_get_jit_exception ();
-    void * das_get_jit_call_or_fastcall ();
-    void * das_get_jit_call_with_cmres ( );
-    void * das_get_jit_invoke_block ( );
-    void * das_get_jit_invoke_block_with_cmres ( );
-    void * das_get_jit_string_builder ();
-    void * das_get_jit_string_builder_temp ();
-    void * das_get_jit_get_global_mnh ();
-    void * das_get_jit_alloc_heap ();
-    void * das_get_jit_alloc_persistent ();
-    void * das_get_jit_free_heap ();
-    void * das_get_jit_free_persistent ();
-    void * das_get_jit_array_lock ();
-    void * das_get_jit_array_unlock ();
-    void * das_get_jit_table_lock ();
-    void * das_get_jit_table_unlock ();
-    void * das_get_jit_array_resize ();
-    void * das_get_jit_table_at ( int32_t baseType, Context * context, LineInfoArg * at );
-    void * das_get_jit_table_erase ( int32_t baseType, Context * context, LineInfoArg * at );
-    void * das_get_jit_table_find ( int32_t baseType, Context * context, LineInfoArg * at );
-    void * das_get_jit_str_cmp ();
-    void * das_get_jit_str_cat ();
-    void * das_get_jit_ast_typedecl ();
-    void * das_get_jit_prologue ();
-    void * das_get_jit_epilogue ();
-    void * das_get_jit_make_block ();
-    void * das_get_jit_ad_by_sid ();
-    void * das_get_jit_debug ();
-    void * das_get_jit_iterator_iterate();
-    void * das_get_jit_iterator_delete();
-    void * das_get_jit_iterator_close();
-    void * das_get_jit_iterator_first();
-    void * das_get_jit_iterator_next();
     void * das_get_builtin_function_address ( Function * fn, Context * context, LineInfoArg * at );
-    void * das_get_jit_debug_enter ();
-    void * das_get_jit_debug_exit ();
-    void * das_get_jit_debug_line ();
 }
